requirements.txt
Django==4.2.11
djangorestframework==3.15.1
httpcore==1.0.5
httpx==0.27.0
openai==1.25.0
pip==21.2.4
pydantic==2.7.1
pydantic_core==2.18.2
setuptools==58.0.4
sniffio==1.3.1
sqlparse==0.5.0
